+++ controllers/modules/upload.js
@@ -117,7 +117,7 @@ const finishExport = (req, res, configs, next) => {
 
 
 // middleware
-const uploadController = (configs) => function uploadController(req, res, next) {
+const parseMultipart = (configs) => function uploadController(req, res, next) {
   req.body.busboySlip = { raw: {}, mes: [] };
   if (!configs.fileSize) configs.fileSize = 25 * 1048576;
   if (!configs.MIME) configs.MIME = ['image/png', 'image/gif', 'image/jpeg', 'image/svg+xml', 'image/x-icon'];
@@ -132,7 +132,7 @@ const uploadController = (configs) => function uploadController(req, res, next) 
 
 // exports
 module.exports = {
-  uploadController,
+  parseMultipart,
   [Symbol.for('UNIT_TEST')]: {
     checkStatus,
     getUploadPath,