+++ controllers/routers/blog.js
@@ -4,49 +4,44 @@ const { PostsModel } = require('../../models/');
 
 
 // controllers
-const blog = { editor: {}, posts: {} };
+const editor = {};
 
-blog.editor.post = {
-  GET: function posts_editor_post_GET(req, res, next) {
+editor.new = {
+  GET: function blog_editor_post_GET(req, res, next) {
     req.session.chest = { post: {} };
     return next();
   },
-  POST: async function posts_editor_post_POST(req, res) {
+  POST: async function blog_editor_post_POST(req, res) {
     await PostsModel.create({ author: req.session.user, ...req.body.post });
     req.flash('info', 'post have been successfully posted!');
     return res.redirect('/blog');
   },
 };
 
-blog.editor.edit = {
-  alias: async function posts_editor_edit_alias(req, res) {
+
+editor.post = {
+  alias: async function blog_editor_edit_alias(req, res) {
     req.session.chest = { post: await PostsModel.findOne(req.params) };
     return res.redirect(`/blog/${req.session.chest.post._id}/edit`);
   },
-  GET: async function posts_editor_edit_GET(req, res, next) {
+  GET: async function blog_editor_edit_GET(req, res, next) {
     if (!req.session.chest) req.session.chest = { post: await PostsModel.findById(_U_.string.readMongoId(req.url)) };
     return next();
   },
-  PATCH: async function posts_editor_edit_PATCH(req, res) {
-    await PostsModel.update({ _id: _U_.string.readMongoId(req.url) }, { $set: req.body.post }, { new: true });
-    req.flash('info', 'post have been successfully updated!');
-    return res.redirect(`/blog/${_U_.string.readMongoId(req.url)}`);
-  },
-  DELETE: async function posts_editor_edit_DELETE(req, res) {
-    await PostsModel.remove({ _id: _U_.string.readMongoId(req.url) });
-    req.flash('info', 'post have been successfully deleted!');
-    return res.redirect('/blog/');
-  },
+
 };
 
-blog.posts.show = {
-  alias: async function posts_posts_show_alias(req, res) {                                                                                          // tofix: revise alias behavior (merged into `postHandler`)
+
+const blog = {};
+
+blog.post = {
+  alias: async function blog_post_alias(req, res) {                                                                                          // tofix: revise alias behavior (merged into `postHandler`)
     const post = await PostsModel.findOne({ ...req.params, 'time._recycled': { $eq: null } });
     if (!post) throw new _U_.error.HttpException(404);
     else req.session.chest = { post };
     return res.redirect(`/blog/${req.session.chest.post.canonical}`);
   },
-  GET: async function posts_posts_show_GET(req, res, next) {                                                                                      // tofix: post not found page
+  GET: async function blog_post_GET(req, res, next) {                                                                                      // tofix: post not found page
     if (!req.session.chest) {
       req.session.chest = {
         post: await PostsModel.findOne({ canonical: req.params[0], 'time._recycled': { $eq: null } }),
@@ -54,12 +49,25 @@ blog.posts.show = {
     }
     return next();
   },
+  PATCH: async function blog_post_PATCH(req, res) {
+    await PostsModel.update({ _id: _U_.string.readMongoId(req.url) }, { $set: req.body.post }, { new: true });
+    req.flash('info', 'post have been successfully updated!');
+    return res.redirect(`/blog/${_U_.string.readMongoId(req.url)}`);
+  },
+  DELETE: async function blog_post_DELETE(req, res) {
+    await PostsModel.remove({ _id: _U_.string.readMongoId(req.url) });
+    req.flash('info', 'post have been successfully deleted!');
+    return res.redirect('/blog/');
+  },
 };
 
-blog.posts.list = {
-  GET: _M_.aggregateFetch('posts'),
+
+blog.list = {
+  GET: function blog_list_GET(req, res, next) {
+    return _U_.express.insertMiddleware(_M_.aggregateFetch('posts'))(req, res, next);
+  },
 };
 
 
 // exports
-module.exports = blog;
+module.exports = { editor, blog };