+++ controllers/routers/page.js
@@ -1,5 +1,5 @@
 const { _M_ } = require('../modules/');
-
+const { _U_ } = require('../utilities/');
 
 // controllers
 const page = {};
@@ -10,17 +10,20 @@ page.search = {
   },
 };
 
+
 page.edit = {
   GET: function page_landing_GET(req, res, next) {
     return next();
   },
 };
 
+
 page.show = {
   GET: function page_landing_GET(req, res, next) {
-    return next();
+    return next(new _U_.error.HttpException(404));
   },
 };
+
 
 page.root = {
   GET: function page_landing_GET(req, res, next) {